+++ backend/src/tiedostot/lisaAineistoService.ts
@@ -1,49 +1,12 @@
-import {
-  AineistoTila,
-  LadattavaTiedosto,
-  LadattavatTiedostot,
-  LisaAineistoParametrit,
-  ListaaLisaAineistoInput,
-} from "hassu-common/graphql/apiModel";
+import { LisaAineistoParametrit, ListaaLisaAineistoInput } from "hassu-common/graphql/apiModel";
 
 import crypto from "crypto";
 import dayjs from "dayjs";
 import { IllegalAccessError, NotFoundError } from "hassu-common/error";
-import { Aineisto, DBProjekti, NahtavillaoloVaihe, NahtavillaoloVaiheJulkaisu } from "../database/model";
-import { fileService } from "../files/fileService";
 import { log } from "../logger";
 import { nyt } from "../util/dateUtil";
-import { jarjestaTiedostot } from "hassu-common/util/jarjestaTiedostot";
 
 class LisaAineistoService {
-  async listaaLisaAineisto(projekti: DBProjekti, params: ListaaLisaAineistoInput): Promise<LadattavatTiedostot> {
-    const nahtavillaolo = findNahtavillaoloVaiheById(projekti, params.nahtavillaoloVaiheId);
-
-    async function adaptLisaAineisto(aineisto: Aineisto): Promise<LadattavaTiedosto> {
-      const { jarjestys, kategoriaId } = aineisto;
-      const nimi = aineisto.nimi;
-      let linkki;
-      if (aineisto.tila == AineistoTila.VALMIS) {
-        if (!aineisto.tiedosto) {
-          const msg = `Virhe lisäaineiston listaamisessa: Aineistolta (nimi: ${nimi}, dokumenttiOid: ${aineisto.dokumenttiOid}) puuttuu tiedosto!`;
-          log.error(msg, { aineisto });
-          throw new Error(msg);
-        }
-        linkki = await fileService.createYllapitoSignedDownloadLink(projekti.oid, aineisto.tiedosto);
-      } else {
-        linkki = "";
-      }
-      return { __typename: "LadattavaTiedosto", nimi, jarjestys, kategoriaId, linkki };
-    }
-
-    const aineistot = (await Promise.all(nahtavillaolo?.aineistoNahtavilla?.map(adaptLisaAineisto) || [])).sort(jarjestaTiedostot) || [];
-    const lisaAineistot = (await Promise.all(nahtavillaolo?.lisaAineisto?.map(adaptLisaAineisto) || [])).sort(jarjestaTiedostot) || [];
-    const aineistopaketti = nahtavillaolo?.aineistopaketti
-      ? await fileService.createYllapitoSignedDownloadLink(projekti.oid, nahtavillaolo?.aineistopaketti)
-      : null;
-    return { __typename: "LadattavatTiedostot", aineistot, lisaAineistot, poistumisPaiva: params.poistumisPaiva, aineistopaketti };
-  }
-
   generateListingParams(oid: string, nahtavillaoloVaiheId: number, salt: string): LisaAineistoParametrit {
     const expires = nyt().add(180, "day").format("YYYY-MM-DD");
     if (!salt) {
@@ -87,22 +50,6 @@ class LisaAineistoService {
       .update([oid, String(params.nahtavillaoloVaiheId), params.poistumisPaiva, salt].join())
       .digest("hex");
   }
-}
-
-function findNahtavillaoloVaiheById(
-  projekti: DBProjekti,
-  nahtavillaoloVaiheId: number
-): NahtavillaoloVaiheJulkaisu | NahtavillaoloVaihe | undefined {
-  let lista: (NahtavillaoloVaiheJulkaisu | NahtavillaoloVaihe | undefined)[] = [];
-  if (projekti.nahtavillaoloVaiheJulkaisut) {
-    lista = lista.concat(projekti.nahtavillaoloVaiheJulkaisut);
-  }
-  if (projekti.nahtavillaoloVaihe) {
-    lista = lista.concat(projekti.nahtavillaoloVaihe);
-  }
-  lista = lista.filter((nahtavillaolo) => nahtavillaolo?.id == nahtavillaoloVaiheId);
-
-  return lista.pop();
 }
 
 export const lisaAineistoService = new LisaAineistoService();
